fix: rename session error classes
