+++ src/domain/sessions/entities/errors/invalid-session-date-expired-error.ts
