+++ src/domain/sessions/entities/errors/invalid-session-date-revoked-error.ts
